master

+++ API/Entities/tswami_gita_scsv.cs
@@ -7,7 +7,7 @@ namespace API.Entities
         public string COMMENT { get; set; }
         public string Chapter { get; set; }
         public string IGS { get; set; }
-        public int NUMBER { get; set; }
+        public double NUMBER { get; set; }
         public string TRANSLITERATION { get; set; }
         public string Gita { get; set; }
         public string Bible_Link { get; set; }
